research/pyproject.toml
[tool.poetry]
name = "mlops-course"
version = "0.1.0"
description = ""
authors = ["Daniil Solovev <dan.solovjew@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pandas = "^2.2.2"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
scikit-learn = "^1.5.0"
matplotlib = "^3.9.0"
mlflow = "^2.13.1"

catboost = "^1.2.5"
[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
black = "^24.4.2"
dvc = {extras = ["s3"], version = "^3.51.0"}


mypy = "^1.10.0"
[tool.poetry.group.docs.dependencies]
quarto = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
extend-include = ["*.ipynb"]

[tool.black]
line_length = 120

[tool.ruff.lint]
select = ["E", "F", "B", "D", "SIM", "I"]
ignore = ["F401", "D211", "D213"]

[tool.mypy]
strict = true
warn_unused_ignores = true
disable_error_code = "import-untyped"